@@ -4,85 +4,39 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node decorate-angular-cli.js && ngcc --properties es2015 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/passport": "^8.2.1",
-    "@nestjs/jwt": "^8.0.0",
-    "@nrwl/angular": "14.0.3",
-    "chart.js": "^3.7.1",
-    "mongoose": "^6.3.1",
+    "mongoose": "^6.8.1",
     "mongoose-autopopulate": "^0.16.1",
     "mongoose-query-parser": "^1.3.1",
-    "moment-timezone": "^0.5.34",
-    "passport": "^0.5.2",
-    "passport-jwt": "^4.0.0",
-    "primeflex": "^3.1.3",
-    "primeng": "^13.3.3",
-    "primeicons": "^5.0.0",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.0.0",
-    "quill": "^1.3.7",
-    "vk-io": "^4.5.2",
-    "zone.js": "~0.11.4"
+    "moment-timezone": "^0.5.40",
+    "tslib": "^2.3.0",
+    "vk-io": "^4.7.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-devkit/architect": "^0.1303.4",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/mongoose": "^9.0.3",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "14.0.3",
-    "@nrwl/cypress": "14.0.3",
-    "@nrwl/eslint-plugin-nx": "14.0.3",
-    "@nrwl/jest": "14.0.3",
-    "@nrwl/linter": "14.0.3",
-    "@nrwl/nest": "14.0.3",
-    "@nrwl/node": "14.0.3",
-    "@nrwl/tao": "^14.0.3",
-    "@nrwl/workspace": "14.0.3",
-    "@types/jest": "27.4.1",
-    "@types/node": "16.11.7",
-    "@types/express": "^4.17.13",
-    "@types/passport": "^1.0.7",
-    "@types/passport-jwt": "^3.0.6",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
-    "class-validator": "^0.13.2",
+    "@nrwl/eslint-plugin-nx": "15.4.2",
+    "@nrwl/jest": "15.4.2",
+    "@nrwl/linter": "15.4.2",
+    "@nrwl/node": "15.4.2",
+    "@nrwl/workspace": "15.4.2",
+    "@types/jest": "28.1.1",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "^5.36.1",
+    "@typescript-eslint/parser": "^5.36.1",
+    "class-validator": "^0.14.0",
     "class-transformer": "^0.5.1",
-    "cypress": "^9.1.0",
-    "eslint": "~8.12.0",
+    "eslint": "~8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "nx": "14.0.3",
-    "prettier": "^2.5.1",
-    "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "~4.6.2"
+    "jest": "28.1.1",
+    "jest-environment-jsdom": "28.1.1",
+    "nx": "15.4.2",
+    "prettier": "^2.6.2",
+    "ts-jest": "28.0.5",
+    "ts-node": "10.9.1",
+    "typescript": "~4.8.2"
   }
 }
